third_party/blink/web_tests/external/wpt/feature-policy/idlharness.window-expected.txt

This is a testharness.js-based test.
PASS idl_test setup
PASS Partial interface Document: original interface defined
PASS Partial interface HTMLIFrameElement: original interface defined
FAIL Stringification of document.featurePolicy assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL FeaturePolicy interface: document.featurePolicy must inherit property "allowsFeature(DOMString, DOMString)" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL FeaturePolicy interface: calling allowsFeature(DOMString, DOMString) on document.featurePolicy with too few arguments must throw TypeError assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL FeaturePolicy interface: document.featurePolicy must inherit property "features()" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL FeaturePolicy interface: document.featurePolicy must inherit property "allowedFeatures()" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL FeaturePolicy interface: document.featurePolicy must inherit property "getAllowlistForFeature(DOMString)" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL FeaturePolicy interface: calling getAllowlistForFeature(DOMString) on document.featurePolicy with too few arguments must throw TypeError assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL FeaturePolicyViolationReportBody interface: existence and properties of interface object assert_own_property: self does not have own property "FeaturePolicyViolationReportBody" expected property "FeaturePolicyViolationReportBody" missing
FAIL FeaturePolicyViolationReportBody interface object length assert_own_property: self does not have own property "FeaturePolicyViolationReportBody" expected property "FeaturePolicyViolationReportBody" missing
FAIL FeaturePolicyViolationReportBody interface object name assert_own_property: self does not have own property "FeaturePolicyViolationReportBody" expected property "FeaturePolicyViolationReportBody" missing
FAIL FeaturePolicyViolationReportBody interface: existence and properties of interface prototype object assert_own_property: self does not have own property "FeaturePolicyViolationReportBody" expected property "FeaturePolicyViolationReportBody" missing
FAIL FeaturePolicyViolationReportBody interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "FeaturePolicyViolationReportBody" expected property "FeaturePolicyViolationReportBody" missing
FAIL FeaturePolicyViolationReportBody interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "FeaturePolicyViolationReportBody" expected property "FeaturePolicyViolationReportBody" missing
FAIL FeaturePolicyViolationReportBody interface: attribute featureId assert_own_property: self does not have own property "FeaturePolicyViolationReportBody" expected property "FeaturePolicyViolationReportBody" missing
FAIL FeaturePolicyViolationReportBody interface: attribute sourceFile assert_own_property: self does not have own property "FeaturePolicyViolationReportBody" expected property "FeaturePolicyViolationReportBody" missing
FAIL FeaturePolicyViolationReportBody interface: attribute lineNumber assert_own_property: self does not have own property "FeaturePolicyViolationReportBody" expected property "FeaturePolicyViolationReportBody" missing
FAIL FeaturePolicyViolationReportBody interface: attribute columnNumber assert_own_property: self does not have own property "FeaturePolicyViolationReportBody" expected property "FeaturePolicyViolationReportBody" missing
FAIL FeaturePolicyViolationReportBody interface: attribute disposition assert_own_property: self does not have own property "FeaturePolicyViolationReportBody" expected property "FeaturePolicyViolationReportBody" missing
FAIL HTMLIFrameElement interface: attribute featurePolicy assert_true: The prototype object must have a property "featurePolicy" expected true got false
FAIL Document interface: attribute featurePolicy assert_true: The prototype object must have a property "featurePolicy" expected true got false
FAIL Document interface: document must inherit property "featurePolicy" with the proper type assert_inherits: property "featurePolicy" not found in prototype chain
Harness: the test ran to completion.
